code_base/backend_api/requirements.txt
fastapi
uvicorn
authlib
httpx
boto3
redis
psycopg2-binary
pymongo
python-dotenv
PyJWT
python-multipart
google-cloud-aiplatform
google-genai
itsdangerous
reportlab
